backend/requirements.txt
anyio==3.6.1
asgiref==3.4.1
attrs==21.4.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==35.0.0
Deprecated==1.2.13
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
phonenumbers==8.12.0
pycparser==2.21
pycryptodome==3.10.4
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.27.1
requests-file==1.5.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
supertokens-python==0.8.1
tldextract==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.1
